# API_documentation/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/mastertables/position": {
      "get": {
        "operationId": "MasterController_getPosition",
        "parameters": [],
        "responses": {
          "201": {
            "description": "List of positions retrieve"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all positions",
        "tags": [
          "Mastertables"
        ]
      },
      "post": {
        "operationId": "MasterController_createPosition",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create Position",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new position",
        "tags": [
          "Mastertables"
        ]
      },
      "patch": {
        "operationId": "MasterController_updatePositionInfo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to update Position Info",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a current position information",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/position/deactivate": {
      "patch": {
        "operationId": "MasterController_deactivatePos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to Deactivate Position",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position has been deactivated"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          },
          "409": {
            "description": "Conflict - Resource deactivated already"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Deactivate a current position",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/position/reactivate": {
      "patch": {
        "operationId": "MasterController_reactivatePos",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to Reactivate Position",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactivatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position has been activated"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          },
          "409": {
            "description": "Conflict - Resource deactivated already"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Reactivate a current Position",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/position/status": {
      "get": {
        "operationId": "MasterController_positionStat",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Position status successfully fetched"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Status of Positions",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/department": {
      "get": {
        "operationId": "MasterController_getDepartment",
        "parameters": [],
        "responses": {
          "201": {
            "description": "List of departments retrieved"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all departments",
        "tags": [
          "Mastertables"
        ]
      },
      "post": {
        "operationId": "MasterController_createDepartment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create Department",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new department",
        "tags": [
          "Mastertables"
        ]
      },
      "patch": {
        "operationId": "MasterController_updateDept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to update department",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDeptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a current department information",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/department/deactivate": {
      "patch": {
        "operationId": "MasterController_deactivateDept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to deactivate department",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department deactivated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          },
          "409": {
            "description": "Conflict - Resource deactivated already"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Deactivate a department",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/department/reactivate": {
      "patch": {
        "operationId": "MasterController_reactivateDept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to activate department",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactivateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department activated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          },
          "409": {
            "description": "Conflict - Resource deactivated already"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Activate a department",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/departments": {
      "get": {
        "operationId": "MasterController_departmentStat",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Department status fetched succussfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get department status",
        "tags": [
          "Mastertables"
        ]
      }
    },
...
  },
  "info": {
    "title": "ABAS v3 API",
    "description": "API documentation for ABAS v3 project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreatePositionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Software Engineer",
            "description": "The name of the position"
          },
          "department_id": {
            "type": "number",
            "example": 2,
            "description": "ID of the department this position belongs to"
          },
          "status": {
            "type": "number",
            "example": "active",
            "description": "active = 1"
          }
        },
        "required": [
          "name",
          "department_id",
          "status"
        ]
      },
      "UpdatePositionDto": {
        "type": "object",
        "properties": {
          "position_id": {
            "type": "number",
            "example": 1,
            "description": "ID of the position you want to update"
          },
          "position_name": {
            "type": "string",
            "example": "New Position name",
            "description": "If you want to update the Position Name"
          },
          "department_id": {
            "type": "number",
            "example": 1,
            "description": "ID of the department you want to update"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "Must be lowercase, if you want to update your Position status"
          }
        },
        "required": [
          "position_id",
          "position_name",
          "department_id",
          "status"
        ]
      },
      "DeactivatePositionDto": {
        "type": "object",
        "properties": {
          "position_id": {
            "type": "number",
            "example": 0,
            "description": "ID of the position you want to deactivate"
          },
          "status": {
            "type": "number",
            "example": "inactive",
            "description": "Must be lowercase, if you want to update your Position status"
          }
        },
        "required": [
          "position_id",
          "status"
        ]
      },
      "ReactivatePositionDto": {
        "type": "object",
        "properties": {
          "position_id": {
            "type": "number",
            "example": 1,
            "description": "ID of the position you want to activate"
          },
          "status": {
            "type": "number",
            "example": "active",
            "description": "Must be lowercase, if you want to update your Position status"
          }
        },
        "required": [
          "position_id",
          "status"
        ]
      },
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Human Resources",
            "description": "The name of the department"
          },
          "division_id": {
            "type": "number",
            "example": 2,
            "description": "ID of the division where the department belongs to"
          },
          "status": {
            "type": "number",
            "example": "active",
            "description": "active = 1"
          }
        },
        "required": [
          "name",
          "division_id",
          "status"
        ]
      },
      "UpdateDeptDto": {
        "type": "object",
        "properties": {
          "department_id": {
            "type": "number",
            "example": 1,
            "description": "ID of the department you want to update"
          },
          "department_name": {
            "type": "string",
            "example": "New Department name",
            "description": "If you want to update the Department Name"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "Must be lowercase, if you want to update your Department status"
          }
        },
        "required": [
          "department_id",
          "department_name",
          "status"
        ]
      },
      "DeactivateDepartmentDto": {
        "type": "object",
        "properties": {
          "department_id": {
            "type": "number",
            "example": 0,
            "description": "ID of the department you want to deactivate"
          },
          "status": {
            "type": "number",
            "example": "inactive",
            "description": "Must be lowercase, if you want to update your Position status"
          }
        },
        "required": [
          "department_id",
          "status"
        ]
      },
      "ReactivateDepartmentDto": {
        "type": "object",
        "properties": {
          "department_id": {
            "type": "number",
            "example": 1,
            "description": "ID of the department you want to activate"
          },
          "status": {
            "type": "number",
            "example": "active",
            "description": "Must be lowercase, if you want to update your Position status"
          }
        },
        "required": [
          "department_id",
          "status"
        ]
      },
...
    }
  }
}